feat(): implement resource views and serializers
