+++ mockserver/api/urls.py
@@ -1,8 +1,3 @@
-from rest_framework import routers
-
-from api.views import MockViewset
-
-router = routers.SimpleRouter()
-router.register('mocks', MockViewset)
+from mocks.urls import router
 
 urlpatterns = router.urls
