+++ mockserver/mocks/serializers.py
@@ -1,8 +1,20 @@
-from mocks.models import Mock
+from mocks.models import Mock, HeaderType, HttpVerb
 from rest_framework import serializers
 
 
 class MockSerializer(serializers.ModelSerializer):
     class Meta:
         model = Mock
+        fields = '__all__'
+
+
+class HeaderTypeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = HeaderType
+        fields = '__all__'
+
+
+class HttpVerbSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = HttpVerb
         fields = '__all__'
