+++ mockserver/mocks/urls.py
@@ -1,6 +1,13 @@
 from django.urls import re_path
+from rest_framework import routers
 
-from mocks.views import fetch_mock
+from mocks.views import fetch_mock, MockViewset, HeaderTypeViewset, HttpVerbViewset
+
+
+router = routers.SimpleRouter()
+router.register('mocks', MockViewset)
+router.register('header-types', HeaderTypeViewset)
+router.register('http-verbs', HttpVerbViewset)
 
 
 urlpatterns = [
